Improve audio output format selection
We try to keep the same format (sample rate and number of channels) for audio output as in the original file.
Because of Qt's API, that implies decoding the first samples to determine the format of the file, then setting a new decoder with the appropriate format.
If that format is not supported by the audio device, we fall back on the default format for that device.